Python/requirements.txt
easygui==0.98.1
numpy==1.18.5
pandas==0.24.2
pyserial==3.4
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
XlsxWriter==1.3.3
